+++ src/nwbinspector/checks/__init__.py
@@ -5,6 +5,7 @@ from ._behavior import (
     check_spatial_series_radians_magnitude,
 )
 from ._ecephys import (
+    check_ascending_spike_times,
     check_electrical_series_dims,
     check_electrical_series_reference_electrodes_table,
     check_negative_spike_times,
@@ -12,6 +13,7 @@ from ._ecephys import (
 )
 from ._general import (
     check_description,
+    check_name_colons,
     check_name_slashes,
 )
 from ._icephys import (
@@ -41,11 +43,13 @@ from ._nwbfile_metadata import (
     check_institution,
     check_keywords,
     check_processing_module_name,
+    check_session_id_no_slashes,
     check_session_start_time_future_date,
     check_session_start_time_old_date,
     check_subject_age,
     check_subject_exists,
     check_subject_id_exists,
+    check_subject_id_no_slashes,
     check_subject_proper_age_range,
     check_subject_sex,
     check_subject_species_exists,
@@ -77,6 +81,7 @@ from ._time_series import (
     check_data_orientation,
     check_missing_unit,
     check_rate_is_not_zero,
+    check_rate_is_positive,
     check_regular_timestamps,
     check_resolution,
     check_timestamp_of_the_first_sample_is_not_negative,
@@ -92,6 +97,7 @@ __all__ = [
     "check_spike_times_not_in_unobserved_interval",
     "check_description",
     "check_name_slashes",
+    "check_name_colons",
     "check_image_series_data_size",
     "check_image_series_external_file_relative",
     "check_image_series_external_file_valid",
@@ -112,9 +118,11 @@ __all__ = [
     "check_experimenter_exists",
     "check_experiment_description",
     "check_subject_id_exists",
+    "check_subject_id_no_slashes",
     "check_subject_species_exists",
     "check_subject_species_form",
     "check_subject_proper_age_range",
+    "check_session_id_no_slashes",
     "check_session_start_time_future_date",
     "check_processing_module_name",
     "check_session_start_time_old_date",
@@ -148,4 +156,6 @@ __all__ = [
     "check_spatial_series_radians_magnitude",
     "check_spatial_series_dims",
     "check_spatial_series_degrees_magnitude",
+    "check_ascending_spike_times",
+    "check_rate_is_positive",
 ]