+++ tests/unit_tests/test_general.py
@@ -1,7 +1,7 @@
 from hdmf.common import DynamicTable, DynamicTableRegion
 
 from nwbinspector import Importance, InspectorMessage
-from nwbinspector.checks import check_description, check_name_slashes
+from nwbinspector.checks import check_description, check_name_colons, check_name_slashes
 
 
 def test_check_name_slashes_pass():
@@ -10,17 +10,48 @@ def test_check_name_slashes_pass():
 
 
 def test_check_name_slashes_fail():
-    """HDMF/PyNWB forbid "/" in the object names. Might need an external file written in MATLAB to test that?"""
-    for x in ["\\"]:
-        table = DynamicTable(name=f"test{x}ing", description="")
-        assert check_name_slashes(neurodata_object=table) == InspectorMessage(
-            message="Object name contains slashes.",
-            importance=Importance.CRITICAL,
-            check_function_name="check_name_slashes",
-            object_type="DynamicTable",
-            object_name=f"test{x}ing",
-            location="/",
-        )
+    # the latest version of HDMF/PyNWB forbids "/" in the object names when creating a new object
+    # so we use in_construct_mode=True to simulate an object that was read from a file
+    table = DynamicTable.__new__(DynamicTable, in_construct_mode=True)
+    table.__init__(name="test/ing", description="")
+    assert check_name_slashes(neurodata_object=table) == InspectorMessage(
+        message="Object name contains slashes.",
+        importance=Importance.CRITICAL,
+        check_function_name="check_name_slashes",
+        object_type="DynamicTable",
+        object_name="test/ing",
+        location="/",
+    )
+
+    table = DynamicTable(name="test\\ing", description="")
+    assert check_name_slashes(neurodata_object=table) == InspectorMessage(
+        message="Object name contains slashes.",
+        importance=Importance.CRITICAL,
+        check_function_name="check_name_slashes",
+        object_type="DynamicTable",
+        object_name="test\\ing",
+        location="/",
+    )
+
+
+def test_check_name_colons_pass():
+    table = DynamicTable(name="test_name", description="")
+    assert check_name_colons(neurodata_object=table) is None
+
+
+def test_check_name_colons_fail():
+    # the latest version of HDMF/PyNWB forbids ":" in the object names when creating a new object
+    # so we use in_construct_mode=True to simulate an object that was read from a file
+    table = DynamicTable.__new__(DynamicTable, in_construct_mode=True)
+    table.__init__(name="test:ing", description="")
+    assert check_name_colons(neurodata_object=table) == InspectorMessage(
+        message="Object name contains colons.",
+        importance=Importance.BEST_PRACTICE_SUGGESTION,
+        check_function_name="check_name_colons",
+        object_type="DynamicTable",
+        object_name="test:ing",
+        location="/",
+    )
 
 
 def test_check_description_pass():