+++ tests/unit_tests/test_time_series.py
@@ -7,6 +7,7 @@ from nwbinspector.checks import (
     check_data_orientation,
     check_missing_unit,
     check_rate_is_not_zero,
+    check_rate_is_positive,
     check_regular_timestamps,
     check_resolution,
     check_timestamp_of_the_first_sample_is_not_negative,
@@ -376,5 +377,39 @@ def test_check_resolution_fail():
         check_function_name="check_resolution",
         object_type="TimeSeries",
         object_name="test",
+        location="/",
+    )
+
+
+def test_check_rate_is_positive_pass():
+    time_series = pynwb.TimeSeries(name="test", unit="test_units", data=np.array([1, 2, 3]), rate=4.0)
+    assert check_rate_is_positive(time_series) is None
+
+
+def test_check_rate_is_positive_none_pass():
+    time_series = pynwb.TimeSeries(
+        name="test", unit="test_units", data=np.array([1, 2, 3]), timestamps=np.array([1, 2, 3])
+    )
+    assert check_rate_is_positive(time_series) is None
+
+
+# @pytest.mark.skipif(
+#    version.parse(pynwb.__version__) >= version.parse("2.5.0"),
+#    reason="pynwb >= 2.5.0 prevents setting negative rates"
+# )
+
+
+def test_check_rate_is_positive_fail():
+    # Use __new__ and in_construct_mode=True to bypass pynwb validation
+    rate = -2.0
+    time_series = pynwb.TimeSeries.__new__(pynwb.TimeSeries, in_construct_mode=True)
+    time_series.__init__(name="TimeSeriesTest", unit="n.a.", data=np.array([1, 2, 3]), rate=rate)
+
+    assert check_rate_is_positive(time_series) == InspectorMessage(
+        message=f"TimeSeriesTest has a negative sampling rate value of {rate}Hz which is not valid.",
+        importance=Importance.CRITICAL,
+        check_function_name="check_rate_is_positive",
+        object_type="TimeSeries",
+        object_name="TimeSeriesTest",
         location="/",
     )